Implemented similar methods in PricedItem
